# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * ReportFlow API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface Client {
  id: number;
  userId: string;
  name: string;
  email: string;
  projectType: string;
  brandColor?: string | null;
  logoUrl?: string | null;
  createdAt: string;
  reportCount: number;
}

export interface CreateClientBody {
  name: string;
  email: string;
  projectType: string;
  brandColor?: string | null;
  logoUrl?: string | null;
}

export interface UpdateClientBody {
  name?: string;
  email?: string;
  projectType?: string;
  brandColor?: string | null;
  logoUrl?: string | null;
}

export interface ReportMetricsData {
  organicTraffic?: number | null;
  paidTraffic?: number | null;
  conversions?: number | null;
  conversionRate?: number | null;
  impressions?: number | null;
  clicks?: number | null;
  ctr?: number | null;
  spend?: number | null;
  roas?: number | null;
  socialFollowers?: number | null;
  socialEngagement?: number | null;
  emailSubscribers?: number | null;
  emailOpenRate?: number | null;
  previousOrganicTraffic?: number | null;
  previousConversions?: number | null;
  previousSpend?: number | null;
  notes?: string | null;
}

export interface Report {
  id: number;
  clientId: number;
  userId: string;
  title: string;
  dateRangeStart: string;
  dateRangeEnd: string;
  data: ReportMetricsData;
  shareToken?: string | null;
  isPublic: boolean;
  clientName: string;
  clientEmail: string;
  clientProjectType: string;
  clientBrandColor?: string | null;
  createdAt: string;
}

export interface CreateReportBody {
  clientId: number;
  title: string;
  dateRangeStart: string;
  dateRangeEnd: string;
  data: ReportMetricsData;
  isPublic?: boolean;
}

export interface UpdateReportBody {
  title?: string;
  dateRangeStart?: string;
  dateRangeEnd?: string;
  data?: ReportMetricsData;
  isPublic?: boolean;
}

export interface DashboardMetrics {
  totalClients: number;
  totalReports: number;
  reportsThisMonth: number;
  sharedReports: number;
  recentReports: Report[];
}

export interface TimeSeriesPoint {
  date: string;
  count: number;
}

export type UserProfilePlan =
  (typeof UserProfilePlan)[keyof typeof UserProfilePlan];

export const UserProfilePlan = {
  free: "free",
  starter: "starter",
  pro: "pro",
} as const;

export interface UserProfile {
  id: number;
  userId: string;
  plan: UserProfilePlan;
  onboardingComplete: boolean;
  createdAt: string;
}

export type BillingCheckoutBodyPlan =
  (typeof BillingCheckoutBodyPlan)[keyof typeof BillingCheckoutBodyPlan];

export const BillingCheckoutBodyPlan = {
  starter: "starter",
  pro: "pro",
} as const;

export interface BillingCheckoutBody {
  plan: BillingCheckoutBodyPlan;
}

export interface BillingCheckoutResponse {
  checkoutUrl: string;
}

export interface BillingPortalResponse {
  portalUrl: string;
}

export type SubscriptionPlan =
  (typeof SubscriptionPlan)[keyof typeof SubscriptionPlan];

export const SubscriptionPlan = {
  starter: "starter",
  pro: "pro",
} as const;

export type SubscriptionStatus =
  (typeof SubscriptionStatus)[keyof typeof SubscriptionStatus];

export const SubscriptionStatus = {
  active: "active",
  on_trial: "on_trial",
  paused: "paused",
  cancelled: "cancelled",
  expired: "expired",
  past_due: "past_due",
  unpaid: "unpaid",
} as const;

export interface Subscription {
  id: number;
  userId: string;
  lsSubscriptionId: string;
  lsCustomerId: string;
  lsVariantId: string;
  lsOrderId?: string | null;
  plan: SubscriptionPlan;
  status: SubscriptionStatus;
  renewsAt?: string | null;
  endsAt?: string | null;
  trialEndsAt?: string | null;
  createdAt: string;
  updatedAt: string;
}

export type SaveUserProfileBodyPlan =
  (typeof SaveUserProfileBodyPlan)[keyof typeof SaveUserProfileBodyPlan];

export const SaveUserProfileBodyPlan = {
  free: "free",
  starter: "starter",
  pro: "pro",
} as const;

export interface SaveUserProfileBody {
  plan: SaveUserProfileBodyPlan;
  email?: string;
  name?: string;
}

export type ListReportsParams = {
  clientId?: number;
};