# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: ReportFlow API specification
servers:
  - url: /api
    description: Base API path
...
paths:
  /billing/checkout:
    post:
      operationId: createBillingCheckout
      tags: [billing]
      summary: Create a Lemon Squeezy checkout URL for a paid plan
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/BillingCheckoutBody"
      responses:
        "200":
          description: Checkout URL
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BillingCheckoutResponse"
...
  /reports:
    get:
      operationId: listReports
      tags: [reports]
      summary: List all reports for the authenticated user
      parameters:
        - name: clientId
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: List of reports
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Report"
    post:
      operationId: createReport
      tags: [reports]
      summary: Create a new report
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateReportBody"
      responses:
        "201":
          description: Report created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Report"
  /reports/{reportId}:
    get:
      operationId: getReport
      tags: [reports]
      summary: Get a single report
      parameters:
        - name: reportId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Report found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Report"
        "404":
          description: Not found
    put:
      operationId: updateReport
      tags: [reports]
      summary: Update a report
      parameters:
        - name: reportId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateReportBody"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Report"
    delete:
      operationId: deleteReport
      tags: [reports]
      summary: Delete a report
      parameters:
        - name: reportId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
...
components:
  schemas:
...
    ReportMetricsData:
      type: object
      properties:
        organicTraffic:
          type: integer
          nullable: true
        paidTraffic:
          type: integer
          nullable: true
        conversions:
          type: integer
          nullable: true
        conversionRate:
          type: number
          nullable: true
        impressions:
          type: integer
          nullable: true
        clicks:
          type: integer
          nullable: true
        ctr:
          type: number
          nullable: true
        spend:
          type: number
          nullable: true
        roas:
          type: number
          nullable: true
        socialFollowers:
          type: integer
          nullable: true
        socialEngagement:
          type: number
          nullable: true
        emailSubscribers:
          type: integer
          nullable: true
        emailOpenRate:
          type: number
          nullable: true
        previousOrganicTraffic:
          type: integer
          nullable: true
        previousConversions:
          type: integer
          nullable: true
        previousSpend:
          type: number
          nullable: true
        notes:
          type: string
          nullable: true
    Report:
      type: object
      properties:
        id:
          type: integer
        clientId:
          type: integer
        userId:
          type: string
        title:
          type: string
        dateRangeStart:
          type: string
          format: date
        dateRangeEnd:
          type: string
          format: date
        data:
          $ref: "#/components/schemas/ReportMetricsData"
        shareToken:
          type: string
          nullable: true
        isPublic:
          type: boolean
        clientName:
          type: string
        clientEmail:
          type: string
        clientProjectType:
          type: string
        clientBrandColor:
          type: string
          nullable: true
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - clientId
        - userId
        - title
        - dateRangeStart
        - dateRangeEnd
        - data
        - isPublic
        - clientName
        - clientEmail
        - clientProjectType
        - createdAt
    CreateReportBody:
      type: object
      properties:
        clientId:
          type: integer
        title:
          type: string
        dateRangeStart:
          type: string
          format: date
        dateRangeEnd:
          type: string
          format: date
        data:
          $ref: "#/components/schemas/ReportMetricsData"
        isPublic:
          type: boolean
      required:
        - clientId
        - title
        - dateRangeStart
        - dateRangeEnd
        - data
    UpdateReportBody:
      type: object
      properties:
        title:
          type: string
        dateRangeStart:
          type: string
          format: date
        dateRangeEnd:
          type: string
          format: date
        data:
          $ref: "#/components/schemas/ReportMetricsData"
        isPublic:
          type: boolean
...
    BillingCheckoutBody:
      type: object
      properties:
        plan:
          type: string
          enum: [starter, pro]
      required:
        - plan
    BillingCheckoutResponse:
      type: object
      properties:
        checkoutUrl:
          type: string
      required:
        - checkoutUrl
...